feat: move raw processor.
