+++ pyfishsense/__init__.py
@@ -2,5 +2,5 @@ from .fish_segmentation_inference import FishSegmentationInference
 from .fish_head_tail_detector import FishHeadTailDetector
 from .image_rectifier import ImageRectifier
 from .laser_detector import LaserDetector
-from .raw_processor import RawProcessor
 from .world_point_handler import WorldPointHandler
+from .image_processors import *
