+++ pyfishsense/image_processors/__init__.py
@@ -1,0 +1,1 @@
+from pyfishsense.image_processors.raw_processor import RawProcessor
