+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "pyFishSense"
+name = "pyfishsense"
 version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
@@ -22,7 +22,7 @@ wheel = "^0.41.3"
 pandas = "^2.1.3"
 
 [tool.poetry.dev-dependencies]
-black = "^22.12.0"
+black = "^23.12.0"
 
 [tool.poetry.group.dev.dependencies]
 ipykernel = "^6.27.1"
